Show reviews

+++ app/Http/Controllers/BookController.php
@@ -8,9 +8,12 @@ use App\Models\Book;
 
 class BookController extends Controller
 {
-    public function index() {
-        return BookResource::collection(Book::all());
-    }
+    public function index()
+{
+    return BookResource::collection(
+        Book::with(['author', 'reviews'])->get()
+    );
+}
 
     public function store(StoreBookRequest $request) {
         $book = Book::create($request->validated());
@@ -26,9 +29,24 @@ class BookController extends Controller
         return BookResource::collection($books);
     }
 
-    public function destroy(Book $book) {
+    // public function destroy(Book $book) {
+    //     $book->delete();
+    //     return response()->json(['message' => 'Book Deleted!']);
+    // }
+
+    public function destroy(Book $book)
+    {
+        if ($book->reviews()->exists()) {
+            return response()->json([
+                'message' => 'This book cannot be deleted because they still have linked reviews.'
+            ], 422);
+        }
+
         $book->delete();
-        return response()->json(['message' => 'Book Deleted!']);
+
+        return response()->json([
+            'message' => 'Book deleted!'
+        ]);
     }
 }
